test_case_reverse_android/requirements.txt
androguard
python-dotenv
Appium-Python-Client 